refactor : genre nav를 만드는 코드를 분리
- genre nav를 만드는 코드를 따로 분리하여 genre-nav.js 파일에 저장
